Added More Functionality & Cleaned
I've added some new functionality to the bot.

The wiki command allows a user to get the first paragraph (or as much as we can show on Discord) of a Wikipedia article of their choosing. It shows a title, a link to the actual wikipedia article (not the api) and the content.

I've also split the commands between several files, as the bot.py file was becoming a little unweildly. So now modifications to each command will be easier, and adds a bit modularity.